+++ hyperion/model/image.py
@@ -39,6 +39,9 @@ class Image(FreezableClass):
         self.lat_min = None
         self.lat_max = None
 
+        self.d_min = None
+        self.d_max = None
+
         self.distance = None
 
         self.pix_area_sr = None
@@ -251,6 +254,34 @@ class Image(FreezableClass):
             self._lat_max = value
         else:
             raise ValueError("lat_max should be a real scalar value")
+
+    @property
+    def d_min(self):
+        """
+        Minimum depth used to define the SEDs (in cm).
+        """
+        return self._d_min
+
+    @d_min.setter
+    def d_min(self, value):
+        if value is None or (np.isscalar(value) and np.isreal(value)):
+            self._d_min = value
+        else:
+            raise ValueError("d_min should be a real scalar value")
+
+    @property
+    def d_max(self):
+        """
+        Maximum depth used to define the SEDs (in cm).
+        """
+        return self._d_max
+
+    @d_max.setter
+    def d_max(self, value):
+        if value is None or (np.isscalar(value) and np.isreal(value)):
+            self._d_max = value
+        else:
+            raise ValueError("d_max should be a real scalar value")
 
     @property
     def distance(self):